Switch to using a Server Extension
I want to provide a button to the Hub Control Panel
when we are running under a JupyterHub. This requires conditional
rendering in some way - either we render the initial HTML as
a template, or setup an API endpoint that JS can use.

This PR templatizes the initial rendering. The following changes
are made:

1. Convert into a Jupyter Server extension. This gives us the
   flexibility we need here.
2. Inherit a handler directly from jupyter_server_proxy, enabling
   us to use the *same* URL for both the initial HTTP rendering
   and the websockify!

However, this means we lose the launcher icon - let's try
mitigate that somehow.